Version 1.13
A bug caused timestamp.vim to make changes in all lines in the |timestamp_modelines| range. Corrected the code so that now changes are made only when there is a timestamp in a particular line. This makes undoing changes a lot nicer, and keeps timestamp-ing out of your way. Also cleaned up some code :)